src/backend/pyproject.toml
[project]
name = "netninja"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.9",
    "aiomysql>=0.2.0",
    "aiosignal==1.3.1",
    "aiosmtplib==3.0.2",
    "aiosqlite==0.20.0",
    "annotated-types==0.7.0",
    "asttokens==2.4.1",
    "attrs==24.2.0",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "colorama==0.4.6",
    "cryptography>=45.0.3",
    "executing==2.1.0",
    "fastapi[standard]>=0.115.12",
    "frozenlist==1.4.1",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "icecream==2.1.3",
    "idna==3.10",
    "iniconfig==2.0.0",
    "markupsafe==3.0.0",
    "multidict==6.1.0",
    "outcome==1.3.0.post0",
    "packaging==24.1",
    "pluggy==1.5.0",
    "pycparser==2.22",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pydantic-settings>=2.9.1",
    "pygments==2.18.0",
    "pysocks==1.7.1",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "requests>=2.32.3",
    "selenium==4.25.0",
    "selenium-async==0.1.0",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "sqlalchemy==2.0.35",
    "sqlmodel>=0.0.24",
    "trio==0.26.2",
    "trio-websocket==0.11.1",
    "typing-extensions==4.12.2",
    "urllib3==2.2.3",
    "websocket-client==1.8.0",
    "wsproto==1.2.0",
    "yarl==1.13.1",
]
